refactor: group all serializers under one directory and separate
